+++ apps/__init__.py
@@ -26,10 +26,6 @@ def register_blueprints(app):
 
 def configure_database(app):
 
-    @app.before_first_request
-    def initialize_database():
-        db.create_all()
-
     @app.teardown_request
     def shutdown_session(exception=None):
         db.session.remove()