+++ apps/authentication/models.py
@@ -9,16 +9,16 @@ from apps import db, login_manager
 
 from apps.authentication.util import hash_pass
 
-class Users(db.Model, UserMixin):
+class users(db.Model, UserMixin):
 
-    __tablename__ = 'Users'
+    __tablename__ = 'users'
 
     id = db.Column(db.Integer, primary_key=True)
     username = db.Column(db.String(64), unique=True)
     email = db.Column(db.String(64), unique=True)
     password = db.Column(db.LargeBinary)
-    recoverEMailToken = db.Column(db.String(100))
-    recoverEMailTimeStamp = db.Column(db.DateTime())
+    recoveremailtoken = db.Column(db.String(100))
+    recoveremailtimestamp = db.Column(db.DateTime())
 
     def __init__(self, **kwargs):
         for property, value in kwargs.items():
@@ -40,11 +40,11 @@ class Users(db.Model, UserMixin):
 
 @login_manager.user_loader
 def user_loader(id):
-    return Users.query.filter_by(id=id).first()
+    return users.query.filter_by(id=id).first()
 
 
 @login_manager.request_loader
 def request_loader(request):
     username = request.form.get('username')
-    user = Users.query.filter_by(username=username).first()
+    user = users.query.filter_by(username=username).first()
     return user if user else None