+++ CHANGELOG.md
@@ -32,7 +32,7 @@
 ### Improvements
 
 - Bump Flask Codebase to [v1.0.7](https://github.com/app-generator/boilerplate-code-flask/releases)
-  - Rename model `User` to `Users` to avoid name conflict with ORACLE DBMS
+  - Rename model `User` to `users` to avoid name conflict with ORACLE DBMS
     - Impacted files: `app/{model.py, util.py, viewws.py}`
 
 ## [1.0.5] 2021-09-16